+++ src/bin/lisp/welcome.lisp
@@ -1,15 +1,3 @@
-(fresh-line)
-(out "LOADables: (Append \".lisp\"!)")(terpri)
-(print '(adjoin aif alist all awhen copy-tree do dolist
-         dotimes ensure-list equality every find-if
-         find group intersect let list max member-if
-         nthcdr prog1 progn queue queue-pop remove-if
-         reverse set-difference set-exclusive-or
-         smoke-test some source stack subseq test
-         test-onerror union unique unless when while
-         with with-queue))
-(terpri)
-
 ; Function called automatically after image load.
 (fn istart ()
   (message "Ready."))